Create a profile page
The profile page now handles all user auth activities, including loggin in and logging out
